OJ-1139: move session2 content into lambda minimal approach
