+++ lambdas/package.json
@@ -6,9 +6,13 @@
   "dependencies": {
       "@aws-lambda-powertools/logger": "^1.5.0",
       "@aws-lambda-powertools/metrics": "^1.5.0",
+      "@aws-lambda-powertools/tracer": "^1.5.1",
       "@aws-sdk/client-dynamodb": "3.150.0",
       "@aws-sdk/client-ssm": "^3.186.0",
       "@aws-sdk/lib-dynamodb": "3.150.0",
+      "@aws-sdk/client-kms": "3.236.0",
+      "@aws-sdk/client-sqs": "3.245.0",
+      "@aws-sdk/credential-providers": "3.245.0",
       "esbuild": "0.14.14",
       "jose": "^4.11.2"
   },
