+++ lambdas/src/common/aws-client-factory.ts
