+++ lambdas/src/common/config/config-service.ts
