+++ lambdas/src/common/config/cri-audit-config.ts
