+++ lambdas/src/common/config/jwt-verification-config.ts
