+++ lambdas/src/common/services/audit-service.ts
