+++ lambdas/src/common/services/session-service.ts
