+++ lambdas/src/common/utils/request-utils.ts
