+++ lambdas/src/services/models/session-request-summary.ts
