+++ lambdas/src/services/session-request-validator.ts
