--- session2/.prettierrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-    semi: true,
-    trailingComma: "all",
-    printWidth: 120,
-    tabWidth: 4,
-};
