--- session2/jest.config.ts
@@ -1,16 +1,0 @@
-/*
- * For a detailed explanation regarding each configuration property and type check, visit:
- * https://jestjs.io/docs/configuration
- */
-
-export default {
-    transform: {
-        "^.+\\.ts?$": "ts-jest",
-    },
-    clearMocks: true,
-    collectCoverage: true,
-    coverageDirectory: "coverage",
-    coverageProvider: "v8",
-    coveragePathIgnorePatterns: ["config.ts", "node_modules/"],
-    testMatch: ["**/tests/**/*.test.ts"],
-};
